Use native extension API for notifications instead of JS Notifications API
JS Notifications API is not very good compared to native extension API. You can
do so much more things and it gives you nearly full control. Also we are getting
the permission for notification while installing the extension. We don't need to
get another permission for JS Notifications API. It can be annoying for the user.